backend/requirements.txt
Django==5.0.2 
djangorestframework==3.14.0 
django-cors-headers==4.3.1 
python-decouple==3.8 
whitenoise==6.6.0 
Pillow==10.0.1 
